refactor(constant): refactor Application
refactor Application :
- 상수 constant로 패키지 위치 수정.
