+++ src/main/java/lotto/dto/LottoResultDto.java
@@ -1,6 +1,6 @@
 package lotto.dto;
 
-import lotto.model.LottoGrade;
+import lotto.constant.LottoGrade;
 
 public record LottoResultDto(
         LottoGrade grade,
