+++ src/main/java/lotto/model/Lotto.java
@@ -3,6 +3,7 @@ package lotto.model;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.constant.LottoNumberRange;
 import lotto.dto.LottoDto;
 import lotto.exception.ExceptionMessage;
